refactor backend concurrency utilities
Split concurrency helpers and isolate administrative user database retries from jobs persistence.\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
